Lab7 without Policie

+++ app/Http/Controllers/CommentController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use App\Models\Comment;
 
@@ -16,17 +18,19 @@ class CommentController extends Controller
         $comment->name = request('name');
         $comment->desc = request('desc');
         $comment->article_id = request('article_id');
-        $comment->user_id = 1;
+        $comment->user_id = Auth::id();
         $comment->save();
         return redirect()->back();
     }
 
     public function edit($id){
         $comment = Comment::findOrFail($id);
+        Gate::authorize('update_comment', $comment);
         return view('comments.update', ['comment'=>$comment]);
     }
 
     public function update(Request $request, Comment $comment){
+        Gate::authorize('update_comment', $comment);
         $request->validate([
             'name'=>'required|min:3',
             'desc'=>'required|max:256'
@@ -39,6 +43,7 @@ class CommentController extends Controller
 
     public function delete($id){
         $comment = Comment::findOrFail($id);
+        Gate::authorize('update_comment', $comment);
         $comment->delete();
         return redirect()->route('article.show', ['article'=>$comment->article_id])->with('status','Delete success');
     }
